Extract password verification to `authentication`
